Tested VhdlTestSelection
HashCalculator.py: Updated get_file_hash method to use os.path.exists to
check if a file is already there. Updated save_hash method to use
os.makedirs(...) to create the directory tree were the checksum file
will live instead of ensure_dir_exists. The ensure_dir_exists was
creating a directory with the name of the file instead of a real file.
Might be a Windows-specific thing...

Renamed RunProject.py to RunVUnit.py

Added first iteration of VhdlTestSelection.py. File is fully functional.
It executes tests when the checksums do not match.